feat: dummy data itemResults 항목 추가
