+++ src/data/dummy.js
@@ -1,0 +1,32 @@
+const itemResults = [
+  {
+    title: '고대맛집',
+    location: '서울특별시 성북구 안암동',
+    thumbnail:
+      'https://user-images.githubusercontent.com/77768091/202692786-76b2d799-8d03-4331-b5de-47009872125b.png',
+    review: '사장님이 친절해요',
+  },
+  {
+    title: '무르무르',
+    location: '서울특별시 성북구 안암동',
+    thumbnail:
+      'https://user-images.githubusercontent.com/77768091/202693110-d5545ca6-0549-4392-aa63-9b9d4b6f18c2.png',
+    review: '비싸요',
+  },
+  {
+    title: '유메노카츠',
+    location: '서울특별시 성북구 하월곡동',
+    thumbnail:
+      'https://user-images.githubusercontent.com/77768091/202693198-8d121c53-94f2-4c50-b4a0-dd857c154f43.png',
+    review: '양이 많아요',
+  },
+  {
+    title: '이정욱',
+    location: '서울특별시 성북구 안암동',
+    thumbnail:
+      'https://user-images.githubusercontent.com/77768091/202694053-4b479dbc-70f3-4dac-bd23-6c2d210e4c81.JPEG',
+    review: '쫄깃쫄깃해요, 웨이팅 있어요!',
+  },
+];
+
+export default itemResults;
