--- src/dummy.js
@@ -1,20 +1,0 @@
-const itemResults = [
-  {
-    title: '고대맛집',
-    location: '서울특별시 성북구 안암동',
-  },
-  {
-    title: '무르무르',
-    location: '서울특별시 성북구 안암동',
-  },
-  {
-    title: '유메노카츠',
-    location: '서울특별시 성북구 하월곡동',
-  },
-  {
-    title: '이정욱',
-    location: '서울특별시 성북구 안암동',
-  },
-];
-
-export default itemResults;
